backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Echo Backend - Time Capsule and Echo Wall Application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "supabase>=2.10.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.15",
    "httpx>=0.27.0",
]
